Add sync with dropbox storage; push conflict copy
Prior to this change, the dropbox storage only downloaded or uploaded
the file which might overwrite the changes name on another device. The
button could trigger several downloads or upload requests, too

This change implement simple sync based on the server modified timestamp
as a version mark. It will upload the local copy with timestamp, e.g.,
`journal 2023-06-02T22/10/21.0.1312.txt` to resolve the issue later on
the desktop platform or later version of the mobile app if it will be an
issue.